+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.22 (pre-release):
+* [Change] Don't allow `KIS` to deploy or move the ground experiments.
 
 # 1.21 (May 16th, 2019):
 * [Fix #89] Portable containers revert to previous content when dropped.